rename to chromaticTextLayer
