+++ packages/sanddance/src/vega-deck.gl/chromatic-text-layer/chromatic-multi-icon-layer/chromatic-multi-icon-layer-vertex.glsl.ts
@@ -18,6 +18,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+//adapted from https://github.com/uber/deck.gl/blob/6.4-release/modules/layers/src/text-layer/multi-icon-layer/multi-icon-layer-vertex.glsl.js
+
 export default `\
 #define SHADER_NAME multi-icon-layer-vertex-shader
