+++ packages/sanddance/src/vega-deck.gl/chromatic-text-layer/tiny-sdf.d.ts
