+++ packages/sanddance/src/vega-deck.gl/exports/types.ts
@@ -33,6 +33,7 @@ import {
     VegaBase
 } from '../base';
 import { CubeLayerDataProps, CubeLayerDefaultProps, CubeLayerProps } from '../cube-layer/cube-layer';
+import { ChromaticTextLayerProps } from '../chromatic-text-layer/chromatic-text-layer';
 import { ViewGlConfig } from '../vega-classes/viewGl';
 
 
@@ -40,6 +41,7 @@ import { ViewGlConfig } from '../vega-classes/viewGl';
 export {
     Axis,
     Base,
+    ChromaticTextLayerProps,
     Cube,
     CubeLayerDataProps,
     CubeLayerDefaultProps,
